pyproject.toml
[project]
name = "opjindal"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "django-environ>=0.12.0",
    "psycopg2>=2.9.10",
]
